// types/Empresa.ts
export type Empresas = {

    EmprCodigo: number;
    EmprNome: string;
    EmprCNPJ: string;
    EmprLink: string;
    EmprLogotipo: Buffer;
    EmprImagemCabecalho: Buffer;
    EmprEndereco: string;
    EmprInicioExpediente: string
    EmprFimExpediente: string
    EmprNumero: string;
    EmprBairro: string;
    EmprCidade: string;
    EmprEstado: string;
    EmprDivisaoSabores: boolean;
    EmprTelefone: string;
    CorSite: string;
  
  };